# pkg/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "LoopWorker API",
    "version": "1.0.0",
    "description": "Work-loop engine REST API. Every response uses one envelope: {success, data|error, timestamp, request_id}. Errors carry a stable machine-readable code plus a message that states the cause and the fix. Credentials: send 'X-API-Key: lwk_...' or 'Authorization: Bearer <token>'. A fresh install prints an ephemeral admin API key once at startup; permanent keys come from LOOPWORKER_API_KEYS or POST /api/v1/auth/keys.",
    "contact": { "name": "LoopWorker self-service docs", "url": "http://localhost:19527/api/v1/openapi.json" }
  },
  "servers": [ { "url": "http://localhost:19527", "description": "local default" } ],
  "security": [ { "ApiKeyHeader": [] }, { "BearerAuth": [] } ],
...
  "paths": {
...
    "/logs": {
      "get": { "tags": ["admin"], "operationId": "adminLogs", "summary": "Observer's in-memory log ring (admin listener only) - always empty in production, because nothing calls Observer.Log",
        "responses": { "200": { "description": "logs (currently always an empty array)" }, "401": { "$ref": "#/components/responses/Unauthorized" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyHeader": { "type": "apiKey", "in": "header", "name": "X-API-Key", "description": "lwk_ prefixed key; stored as a SHA-256 digest" },
      "BearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "lwt", "description": "HMAC-signed token from POST /api/v1/auth/token" }
    },
...
    "schemas": {
...
      "ErrorCode": {
        "type": "string",
        "description": "Stable machine-readable error code. Never changes meaning.",
        "enum": ["INVALID_REQUEST", "UNKNOWN_FIELD", "MALFORMED_JSON", "INPUT_ENCODING_INVALID", "REQUEST_TOO_LARGE",
          "UNAUTHORIZED", "TOKEN_INVALID", "TOKEN_EXPIRED", "FORBIDDEN", "TASK_NOT_FOUND", "TASK_ALREADY_EXISTS",
          "TASK_STATE_CONFLICT", "WORKFLOW_NOT_FOUND", "WORKER_NOT_FOUND", "PLUGIN_NOT_FOUND", "DEPENDENCY_CYCLE",
          "DEPENDENCY_INVALID", "GRAPH_TOO_LARGE", "RATE_LIMITED", "STREAM_LIMIT_REACHED", "QUEUE_FULL", "NOT_FOUND",
          "METHOD_NOT_ALLOWED", "STREAMING_UNSUPPORTED", "SERVICE_UNAVAILABLE", "INTERNAL_ERROR", "TIMEOUT"]
      },
      "ErrorBody": {
        "type": "object",
        "required": ["code", "message", "request_id"],
        "properties": {
          "code": { "$ref": "#/components/schemas/ErrorCode" },
          "message": { "type": "string", "description": "Cause first, then 'Fix: ' with the repair step" },
          "details": { "type": "object", "additionalProperties": true },
          "request_id": { "type": "string" }
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": ["success", "error", "timestamp", "request_id"],
        "properties": {
          "success": { "type": "boolean", "enum": [false] },
          "error": { "$ref": "#/components/schemas/ErrorBody" },
          "timestamp": { "type": "string", "format": "date-time" },
          "request_id": { "type": "string" }
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": { "description": "401 UNAUTHORIZED / TOKEN_INVALID / TOKEN_EXPIRED", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
...
    }
  }
}